Namespace, global app, fold-in Utils.cpp
