+++ LBFRegressor.cpp
@@ -10,23 +10,25 @@
 using namespace std;
 using namespace cv;
 
+namespace FaceAlignment {
+
 struct feature_node ** LBFRegressor::DeriveBinaryFeat(
                                     const RandomForest& randf,
                                     const vector<Mat_<uchar> >& images,
                                     const vector<Mat_<double> >& current_shapes,
                                     const vector<BoundingBox> & bounding_boxs){
-    
+
     // initilaize the memory for binfeatures
     struct feature_node **binfeatures;
     binfeatures = new struct feature_node* [images.size()];
     for (int i=0;i<images.size();i++){
          binfeatures[i] = new struct feature_node[randf.max_numtrees_*randf.num_landmark_+1];
     }
-    
+
 //    int bincode;
 //    int ind;
 //    int leafnode_per_tree = pow(2,(randf.max_depth_-1));
-    
+
     Mat_<double> rotation;
     double scale;
 
@@ -44,7 +46,7 @@ struct feature_node ** LBFRegressor::DeriveBinaryFeat(
 //                binfeatures[i][ind].index = leafnode_per_tree * ind + bincode;
 //                binfeatures[i][ind].value = 1;
 //            }
-            
+
         }
         binfeatures[i][randf.num_landmark_ * randf.max_numtrees_].index = -1;
         binfeatures[i][randf.num_landmark_ * randf.max_numtrees_].value = -1;
@@ -61,7 +63,7 @@ void LBFRegressor::GetCodefromRandomForest(struct feature_node *binfeature,
                                            const BoundingBox& bounding_box,
                                            const Mat_<double>& rotation,
                                            const double scale){
-    
+
     int leafnode_per_tree = pow(2,rand_forest[0].max_depth_-1);
     int landmark_x = shape(rand_forest[0].landmarkID_,0);
     int landmark_y = shape(rand_forest[0].landmarkID_,1);
@@ -74,7 +76,7 @@ void LBFRegressor::GetCodefromRandomForest(struct feature_node *binfeature,
             double y1 = rand_forest[iter].nodes_[currnode].feat[1];
             double x2 = rand_forest[iter].nodes_[currnode].feat[2];
             double y2 = rand_forest[iter].nodes_[currnode].feat[3];
-            
+
             double project_x1 = rotation(0,0) * x1 + rotation(0,1) * y1;
             double project_y1 = rotation(1,0) * x1 + rotation(1,1) * y1;
             project_x1 = scale * project_x1 * bounding_box.width / 2.0;
@@ -83,7 +85,7 @@ void LBFRegressor::GetCodefromRandomForest(struct feature_node *binfeature,
             int real_y1 = (int)project_y1 + landmark_y;
             real_x1 = max(0,min(real_x1,image.cols-1));
             real_y1 = max(0,min(real_y1,image.rows-1));
-            
+
             double project_x2 = rotation(0,0) * x2 + rotation(0,1) * y2;
             double project_y2 = rotation(1,0) * x2 + rotation(1,1) * y2;
             project_x2 = scale * project_x2 * bounding_box.width / 2.0;
@@ -92,7 +94,7 @@ void LBFRegressor::GetCodefromRandomForest(struct feature_node *binfeature,
             int real_y2 = (int)project_y2 + landmark_y;
             real_x2 = max(0,min(real_x2,image.cols-1));
             real_y2 = max(0,min(real_y2,image.rows-1));
-            
+
             int pdf = (int)(image(real_y1,real_x1))-(int)(image(real_y2,real_x2));
             if (pdf < rand_forest[iter].nodes_[currnode].thresh){
                 currnode =rand_forest[iter].nodes_[currnode].cnodes[0];
@@ -104,7 +106,7 @@ void LBFRegressor::GetCodefromRandomForest(struct feature_node *binfeature,
         }
         binfeature[index+iter].index = leafnode_per_tree*(index+iter)+bincode;
         binfeature[index+iter].value = 1;
-        
+
     }
 }
 int  LBFRegressor::GetCodefromTree(const Tree& tree,
@@ -120,7 +122,7 @@ int  LBFRegressor::GetCodefromTree(const Tree& tree,
         double y1 = tree.nodes_[currnode].feat[1];
         double x2 = tree.nodes_[currnode].feat[2];
         double y2 = tree.nodes_[currnode].feat[3];
-        
+
         double project_x1 = rotation(0,0) * x1 + rotation(0,1) * y1;
         double project_y1 = rotation(1,0) * x1 + rotation(1,1) * y1;
         project_x1 = scale * project_x1 * bounding_box.width / 2.0;
@@ -129,7 +131,7 @@ int  LBFRegressor::GetCodefromTree(const Tree& tree,
         int real_y1 = project_y1 + shape(tree.landmarkID_,1);
         real_x1 = max(0,min(real_x1,image.cols-1));
         real_y1 = max(0,min(real_y1,image.rows-1));
-        
+
         double project_x2 = rotation(0,0) * x2 + rotation(0,1) * y2;
         double project_y2 = rotation(1,0) * x2 + rotation(1,1) * y2;
         project_x2 = scale * project_x2 * bounding_box.width / 2.0;
@@ -138,7 +140,7 @@ int  LBFRegressor::GetCodefromTree(const Tree& tree,
         int real_y2 = project_y2 + shape(tree.landmarkID_,1);
         real_x2 = max(0,min(real_x2,image.cols-1));
         real_y2 = max(0,min(real_y2,image.rows-1));
-        
+
         int pdf = (int)(image(real_y1,real_x1))-(int)(image(real_y2,real_x2));
         if (pdf < tree.nodes_[currnode].thresh){
             currnode =tree.nodes_[currnode].cnodes[0];
@@ -169,7 +171,7 @@ void LBFRegressor::GlobalRegression(struct feature_node **binfeatures,
     prob -> n = num_feature;
     prob -> x = binfeatures;
     prob -> bias = -1;
-    
+
     // construct the parameter
     struct parameter* param = new struct parameter;
     param-> solver_type = L2R_L2LOSS_SVR_DUAL;
@@ -178,11 +180,11 @@ void LBFRegressor::GlobalRegression(struct feature_node **binfeatures,
     param->p = 0;
     param->eps = 0.0001;
     //param->eps = 0.001;
-    
+
     // initialize the y
     int num_residual = shapes_residual[0].rows*2;
     double** yy = new double*[num_residual];
-    
+
     for (int i=0;i<num_residual;i++){
         yy[i] = new double[num_train_sample];
     }
@@ -196,7 +198,7 @@ void LBFRegressor::GlobalRegression(struct feature_node **binfeatures,
             }
         }
     }
-    
+
     //train
     models.clear();
     models.resize(num_residual);
@@ -234,7 +236,7 @@ void LBFRegressor::GlobalRegression(struct feature_node **binfeatures,
         transpose(rotation,rotation);
         deltashape_bar1 = scale * deltashape_bar * rotation;
         current_shapes[i] = ReProjectShape((ProjectShape(current_shapes[i],bounding_boxs[i])+deltashape_bar1),bounding_boxs[i]);
-        
+
         //updata shapes_residual
        // shapes_residual[i] = shapes_residual[i] - deltashape_bar;
     }
@@ -276,47 +278,47 @@ void LBFRegressor::GlobalPrediction(struct feature_node** binfeatures,
 void LBFRegressor::Train(const vector<Mat_<uchar> >& images,
                          const vector<Mat_<double> >& ground_truth_shapes,
                          const vector<BoundingBox> & bounding_boxs){
-    
+
     // data augmentation and multiple initialization
     vector<Mat_<uchar> > augmented_images;
     vector<BoundingBox> augmented_bounding_boxs;
     vector<Mat_<double> > augmented_ground_truth_shapes;
     vector<Mat_<double> > current_shapes;
-    
+
     RNG random_generator(getTickCount());
     for(int i = 0;i < images.size();i++){
-        for(int j = 0;j < global_params.initial_num;j++){
+        for(int j = 0;j < global.config.initial_num;j++){
             int index = 0;
             do{
                 // index = (i+j+1) % (images.size());
                 index = random_generator.uniform(0, (int)images.size());
             }while(index == i);
-            
+
             // 1. Select ground truth shapes of other images as initial shapes
             augmented_images.push_back(images[i]);
             augmented_ground_truth_shapes.push_back(ground_truth_shapes[i]);
             augmented_bounding_boxs.push_back(bounding_boxs[i]);
-            
+
             // 2. Project current shape to bounding box of ground truth shapes
             Mat_<double> temp = ProjectShape(ground_truth_shapes[index], bounding_boxs[index]);
             temp = ReProjectShape(temp, bounding_boxs[i]);
             current_shapes.push_back(temp);
         }
     }
-    
+
     // get mean shape from training shapes(only origin train images)
     mean_shape_ = GetMeanShape(ground_truth_shapes,bounding_boxs);
     cout << mean_shape_<<endl;
     // train random forest
-    int num_feature = global_params.landmark_num * global_params.max_numtrees * pow(2,(global_params.max_depth-1));
+    int num_feature = global.config.landmark_num * global.config.max_numtrees * pow(2,(global.config.max_depth-1));
     int num_train_sample = (int)augmented_images.size();
 
     double t0 =(double)cvGetTickCount();
-    for (int stage = 0; stage < global_params.max_numstage; stage++){
+    for (int stage = 0; stage < global.config.max_numstage; stage++){
         double t1 =(double)cvGetTickCount();
         GetShapeResidual(augmented_ground_truth_shapes,current_shapes,augmented_bounding_boxs,
                          mean_shape_,shapes_residual_);
-        
+
         cout << "train random forest of "<< stage <<" stage" <<endl;
         RandomForest_[stage].Train(augmented_images,augmented_ground_truth_shapes, current_shapes, augmented_bounding_boxs, mean_shape_, shapes_residual_, stage);
         double t2 = (double)cvGetTickCount();
@@ -327,24 +329,24 @@ void LBFRegressor::Train(const vector<Mat_<uchar> >& images,
         binfeatures = DeriveBinaryFeat(RandomForest_[stage], augmented_images, current_shapes, augmented_bounding_boxs);
         double t3 = (double)cvGetTickCount();
         cout << "derive binary features of "<< stage<<" stage has been trained, cost "<< (t3-t2)/((double)cvGetTickFrequency()*1000*1000) <<" s"<<endl<<endl;
-        
+
         cout << "learn global linear regression given binary feature" << endl;
         GlobalRegression(binfeatures, shapes_residual_, current_shapes, augmented_bounding_boxs, mean_shape_, Models_[stage], num_feature, num_train_sample, stage);
         ReleaseFeatureSpace(binfeatures,(int)augmented_images.size());
-        
+
         //calculate the error
         double MRSE_sum = 0;
         for (int i =0; i<current_shapes.size();i++){
             MRSE_sum += CalculateError(augmented_ground_truth_shapes[i], current_shapes[i]);
         }
         cout <<"stage "<<stage<<", error: "<<MRSE_sum/current_shapes.size()<<endl;
-        
+
         //calculate the remaining time
         double t4 = (double)cvGetTickCount();
         cout << "the linear model of "<< stage<<" stage has been trained, cost "<< (t4-t3)/((double)cvGetTickFrequency()*1000*1000) <<" s"<<endl<<endl;
 
         cout << "the "<<stage<<" has completed, cost "<<(t4-t0)/((double)cvGetTickFrequency()*1000*1000) <<" s"<<endl;
-        cout << "Remaining time is about "<< (t4-t0)/((double)cvGetTickFrequency()*1000*1000*(stage+1))*(global_params.max_numstage-stage-1)<< "s"<<endl<<endl;
+        cout << "Remaining time is about "<< (t4-t0)/((double)cvGetTickFrequency()*1000*1000*(stage+1))*(global.config.max_numstage-stage-1)<< "s"<<endl<<endl;
     }
 }
 void LBFRegressor::ReleaseFeatureSpace(struct feature_node ** binfeatures,
@@ -359,7 +361,7 @@ vector<Mat_<double> > LBFRegressor::Predict(const vector<Mat_<uchar> >& images,
                                     const vector<BoundingBox>& bounding_boxs,
                                     const vector<Mat_<double> >& ground_truth_shapes,
                                     int initial_num){
-    
+
     vector<Mat_<double> > current_shapes;
 
     for (int i=0; i<images.size();i++){
@@ -373,12 +375,12 @@ vector<Mat_<double> > LBFRegressor::Predict(const vector<Mat_<uchar> >& images,
     cout <<"mean shape "<<", error: "<<MRSE_sum/current_shapes.size()<<endl;
 
     int stage1 =0;
-    for ( int stage = 0; stage < global_params.max_numstage; stage++){
-        if(stage<global_params.max_numstage){
+    for ( int stage = 0; stage < global.config.max_numstage; stage++){
+        if(stage<global.config.max_numstage){
             stage1 = stage;
         }
         else{
-            stage1 = global_params.max_numstage-1;
+            stage1 = global.config.max_numstage-1;
         }
         struct feature_node ** binfeatures ;
         binfeatures = DeriveBinaryFeat(RandomForest_[stage1],images,current_shapes,bounding_boxs);
@@ -390,7 +392,7 @@ vector<Mat_<double> > LBFRegressor::Predict(const vector<Mat_<uchar> >& images,
         }
         cout <<"stage "<<stage<<", error: "<<MRSE_sum/current_shapes.size()<<endl;
     }
-    
+
     return current_shapes;
 }
 
@@ -405,27 +407,27 @@ Mat_<double>  LBFRegressor::Predict(const cv::Mat_<uchar>& image,
     images.push_back(image);
     bounding_boxs.push_back(bounding_box);
     current_shapes.push_back(ReProjectShape(mean_shape_, bounding_box));
-   
+
 //    Mat img = imread("/Users/lequan/workspace/LBF/Datasets/lfpw/testset/image_0078.png");
 //    // draw result :: red
-//    for(int j = 0;j < global_params.landmark_num;j++){
+//    for(int j = 0;j < global.config.landmark_num;j++){
 //        circle(img,Point2d(current_shapes[0](j,0),current_shapes[0](j,1)),1,Scalar(255,255,255),-1,8,0);
 //    }
 //    imshow("result", img);
 //    waitKey(0);
 //    string name = "example mean.jpg";
 //    imwrite(name,img);
-    
-    
-    for ( int stage = 0; stage < global_params.max_numstage; stage++){
+
+
+    for ( int stage = 0; stage < global.config.max_numstage; stage++){
         struct feature_node ** binfeatures ;
         binfeatures = DeriveBinaryFeat(RandomForest_[stage],images,current_shapes,bounding_boxs);
         GlobalPrediction(binfeatures, current_shapes,bounding_boxs,stage);
         ReleaseFeatureSpace(binfeatures, images.size());
-        
+
 //        Mat image = imread("/Users/lequan/workspace/LBF/Datasets/afw/image_0078.png");
 //        // draw result :: red
-//        for(int j = 0;j < global_params.landmark_num;j++){
+//        for(int j = 0;j < global.config.landmark_num;j++){
 //            circle(image,Point2d(current_shapes[0](j,0),current_shapes[0](j,1)),1,Scalar(255,255,255),-1,8,0);
 //        }
 //        imshow("result", image);
@@ -447,7 +449,7 @@ void LBFRegressor::Save(string path){
     fout.close();
     cout << "End" << endl;
 
-    
+
 }
 
 void LBFRegressor::Load(string path){
@@ -460,33 +462,33 @@ void LBFRegressor::Load(string path){
     cout << "End"<<endl;
 }
 void  LBFRegressor::WriteGlobalParam(ofstream& fout){
-    fout << global_params.bagging_overlap << endl;
-    fout << global_params.max_numtrees << endl;
-    fout << global_params.max_depth << endl;
-    fout << global_params.max_numthreshs << endl;
-    fout << global_params.landmark_num << endl;
-    fout << global_params.initial_num << endl;
-    fout << global_params.max_numstage << endl;
-    
-    for (int i = 0; i< global_params.max_numstage; i++){
-        fout << global_params.max_radio_radius[i] << " ";
-        
+    fout << global.config.bagging_overlap << endl;
+    fout << global.config.max_numtrees << endl;
+    fout << global.config.max_depth << endl;
+    fout << global.config.max_numthreshs << endl;
+    fout << global.config.landmark_num << endl;
+    fout << global.config.initial_num << endl;
+    fout << global.config.max_numstage << endl;
+
+    for (int i = 0; i< global.config.max_numstage; i++){
+        fout << global.config.max_radio_radius[i] << " ";
+
     }
     fout << endl;
-    
-    for (int i = 0; i < global_params.max_numstage; i++){
-        fout << global_params.max_numfeats[i] << " ";
+
+    for (int i = 0; i < global.config.max_numstage; i++){
+        fout << global.config.max_numfeats[i] << " ";
     }
     fout << endl;
 }
 void  LBFRegressor::WriteRegressor(ofstream& fout){
-    for(int i = 0;i < global_params.landmark_num;i++){
+    for(int i = 0;i < global.config.landmark_num;i++){
         fout << mean_shape_(i,0)<<" "<< mean_shape_(i,1)<<" ";
     }
     fout<<endl;
     ofstream fout_reg;
-    fout_reg.open(modelPath + "/Regressor.model",ios::binary);
-    for (int i=0; i < global_params.max_numstage; i++ ){
+    fout_reg.open(global.config.model_path + "/Regressor.model",ios::binary);
+    for (int i=0; i < global.config.max_numstage; i++ ){
         RandomForest_[i].Write(fout);
         fout << Models_[i].size()<< endl;
         for (int j=0; j<Models_[i].size();j++){
@@ -496,31 +498,31 @@ void  LBFRegressor::WriteRegressor(ofstream& fout){
     fout_reg.close();
 }
 void  LBFRegressor::ReadGlobalParam(ifstream& fin){
-    fin >> global_params.bagging_overlap;
-    fin >> global_params.max_numtrees;
-    fin >> global_params.max_depth;
-    fin >> global_params.max_numthreshs;
-    fin >> global_params.landmark_num;
-    fin >> global_params.initial_num;
-    fin >> global_params.max_numstage;
-    
-    for (int i = 0; i< global_params.max_numstage; i++){
-        fin >> global_params.max_radio_radius[i];
+    fin >> global.config.bagging_overlap;
+    fin >> global.config.max_numtrees;
+    fin >> global.config.max_depth;
+    fin >> global.config.max_numthreshs;
+    fin >> global.config.landmark_num;
+    fin >> global.config.initial_num;
+    fin >> global.config.max_numstage;
+
+    for (int i = 0; i< global.config.max_numstage; i++){
+        fin >> global.config.max_radio_radius[i];
     }
-    
-    for (int i = 0; i < global_params.max_numstage; i++){
-        fin >> global_params.max_numfeats[i];
+
+    for (int i = 0; i < global.config.max_numstage; i++){
+        fin >> global.config.max_numfeats[i];
     }
 }
 
 void LBFRegressor::ReadRegressor(ifstream& fin){
-    mean_shape_ = Mat::zeros(global_params.landmark_num,2,CV_64FC1);
-    for(int i = 0;i < global_params.landmark_num;i++){
+    mean_shape_ = Mat::zeros(global.config.landmark_num,2,CV_64FC1);
+    for(int i = 0;i < global.config.landmark_num;i++){
         fin >> mean_shape_(i,0) >> mean_shape_(i,1);
     }
     ifstream fin_reg;
-    fin_reg.open(modelPath + "/Regressor.model",ios::binary);
-    for (int i=0; i < global_params.max_numstage; i++ ){
+    fin_reg.open(global.config.model_path + "/Regressor.model",ios::binary);
+    for (int i=0; i < global.config.max_numstage; i++ ){
         RandomForest_[i].Read(fin);
         int num =0;
         fin >> num;
@@ -532,3 +534,4 @@ void LBFRegressor::ReadRegressor(ifstream& fin){
     fin_reg.close();
 }
 
+}
