+++ TrainModel.cpp
@@ -8,6 +8,9 @@
 #include "LBFRegressor.h"
 using namespace std;
 using namespace cv;
+
+namespace FaceAlignment {
+
 void LoadCofwTrainData(vector<Mat_<uchar> >& images,
                        vector<Mat_<double> >& ground_truth_shapes,
                        vector<BoundingBox>& bounding_boxs);
@@ -19,9 +22,9 @@ void TrainModel(vector<string> trainDataName){
     for(int i=0;i<trainDataName.size();i++){
         string path;
         if(trainDataName[i]=="helen"||trainDataName[i]=="lfpw")
-            path = dataPath + trainDataName[i] + "/trainset/Path_Images.txt";
+            path = global.config.data_path + trainDataName[i] + "/trainset/Path_Images.txt";
         else
-            path = dataPath + trainDataName[i] + "/Path_Images.txt";
+            path = global.config.data_path + trainDataName[i] + "/Path_Images.txt";
 
        // LoadData(path, images, ground_truth_shapes, bounding_boxs);
           LoadOpencvBbxData(path, images, ground_truth_shapes, bounding_boxs);
@@ -29,8 +32,7 @@ void TrainModel(vector<string> trainDataName){
 
     LBFRegressor regressor;
     regressor.Train(images,ground_truth_shapes,bounding_boxs);
-    regressor.Save(modelPath+"LBF.model");
+    regressor.Save(global.config.model_path + "LBF.model");
     return;
 }
-
-
+}
